+++ mocks/src/cosmos/mod.rs
@@ -1,6 +1,8 @@
 mod app;
 mod builder;
 mod helpers;
+mod manual;
+mod runner;
 
 pub use app::*;
 pub use builder::*;
