+++ mocks/src/lib.rs
@@ -6,9 +6,6 @@ pub mod setup;
 #[cfg(feature = "native")]
 pub mod sovereign;
 
-#[cfg(feature = "native")]
-pub mod configs;
-
 pub mod utils;
 
 #[cfg(all(test, feature = "native"))]
