+++ mocks/src/sovereign/mod.rs
@@ -1,9 +1,9 @@
-mod config;
 mod da_service;
+mod manual;
 mod rollup;
+mod runner;
 mod runtime;
 
-pub use config::*;
 pub use da_service::*;
 pub use rollup::*;
 pub use runtime::*;
