+++ mocks/src/utils/mod.rs
@@ -1,5 +1,10 @@
+#[cfg(feature = "native")]
+mod configs;
+
 use std::time::Duration;
 
+#[cfg(feature = "native")]
+pub use configs::*;
 use tokio::time::sleep;
 
 /// Waits for the mock chains to generate a few blocks.
